refactor(workflow): split Workflow.kt by runtime responsibility — Epic 4.2
